pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "terminal-bridge-v2"
version = "0.2.0"
description = "Universal CLI LLM remote control + monitoring + human intervention"
requires-python = ">=3.9"
license = {text = "MIT"}

[project.optional-dependencies]
windows = ["pywinpty>=0.5.4"]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[project.scripts]
tb2 = "tb2.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = ["e2e: end-to-end tests requiring real tmux"]
